Uređeni js za search (treba doradit), darkmode uređen malo, etc.

+++ src/main/java/com/project/Managmently/controllers/authRestricted/AccountController.java
@@ -3,11 +3,13 @@ package com.project.Managmently.controllers.authRestricted;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.project.Managmently.classes.User;
@@ -18,6 +20,9 @@ public class AccountController {
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private BCryptPasswordEncoder passwordEncoder;
     
     @GetMapping("/account")
     public String getContactInfoPage(Model model) {
@@ -45,13 +50,19 @@ public class AccountController {
     }
 
     @PostMapping("/deleteUser")
-    public String deleteUser(@ModelAttribute User user, RedirectAttributes redirectAttributes) {
+    public String deleteUser(@ModelAttribute User user, @RequestParam("password") String password, RedirectAttributes redirectAttributes) {
 
-        try {
-            userRepository.deleteUser(user);
-            redirectAttributes.addFlashAttribute("successMessage", "Account successfully deleted.");
+        String encryptedPassword = passwordEncoder.encode(password);
+        if(encryptedPassword == user.getPassword()) {
+            try {
+                userRepository.deleteUser(user);
+                redirectAttributes.addFlashAttribute("successMessage", "Account successfully deleted.");
 
-        } catch (Exception e) {
+            } catch (Exception e) {
+                redirectAttributes.addFlashAttribute("errorMessage", "There was an issue with deleting your account. Please try again.");
+                return "redirect:/account";
+            }
+        } else {
             redirectAttributes.addFlashAttribute("errorMessage", "There was an issue with deleting your account. Please try again.");
             return "redirect:/account";
         }
